Add golang server, docker image build
This change packages up the site into a Docker container, with an HTML
static fileserver implementation.

This is not as simple as pushing to an external provider, but this
allows for extension of the fileserver (adding metrics, etc.) as well as
being deployable quickly (already have infra to serve from Docker
containers).